Intellion_core/requirements.txt
pandas
requests
python-dotenv
openai
catboost
scikit-learn
beautifulsoup4
lxml
nltk
tqdm
